+++ Changes.md
@@ -17,6 +17,7 @@ API
 - View : Added DisplayTransform add-on class which can be used to add colourspace management to any View.
 - ViewportGadget : A post-process shader can now be applied to any layer, not just the main one.
 - SceneGadget : Added `setLayer()` and `getLayer()` methods, which allow the destination `Gadget::Layer` to be specified.
+- TestCase : Added `ignoreMessage()` method, to register messages that should not be treated as test failures.
 
 Breaking Changes
 ----------------